testit tehty

+++ controllers/blogs.js
@@ -1,21 +1,32 @@
 const blogsRouter = require('express').Router()
 const Blog = require('./../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes === undefined ? 0 : body.likes
+        likes: body.likes === undefined ? 0 : body.likes,
+        user: user._id
     })
     const savedBlog = await blog.save()
-    response.json(savedBlog.toJSON())
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog.toJSON())
 })
 
 blogsRouter.get('/:id', async (request, response) => {
@@ -33,17 +44,29 @@ blogsRouter.put('/:id', async (request, response) => {
     const blog2 = {
         title: body.title || blog1.title,
         author: body.author || blog1.author,
-        url: body.url || blog1.url,
-        likes: body.likes || 0
+        url: body.url || blog1.url,
+        likes: body.likes || 0
     }
-    const update = await Blog.findByIdAndUpdate(request.params.id, blog2, {new: true})
+    const update = await Blog.findByIdAndUpdate(request.params.id, blog2, { new: true })
     response.json(update.toJSON())
 
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: "Token missing or invalid" })
+    }
+
+    const user = await User.findOne({ username: decodedToken.username })
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).end()
+    }
 })
 
 module.exports = blogsRouter
